migrate to latest version

+++ src/orchestrator/transaction/datastore/base-in-memory-storage.ts
@@ -1,6 +1,6 @@
-import { TransactionCheckpoint } from "../distributed-transaction"
-import { TransactionModelOptions } from "../types"
-import { DistributedTransactionStorage } from "./abstract-storage"
+import { TransactionCheckpoint } from '../distributed-transaction'
+import { TransactionOptions, TransactionState } from '../types'
+import { DistributedTransactionStorage } from './abstract-storage'
 
 // eslint-disable-next-line max-len
 export class BaseInMemoryDistributedTransactionStorage extends DistributedTransactionStorage {
@@ -11,7 +11,7 @@ export class BaseInMemoryDistributedTransactionStorage extends DistributedTransa
     this.storage = new Map()
   }
 
-  async get(key: string): Promise<TransactionCheckpoint | undefined> {
+  async get(key: string, options?: TransactionOptions): Promise<TransactionCheckpoint | undefined> {
     return this.storage.get(key)
   }
 
@@ -22,16 +22,19 @@ export class BaseInMemoryDistributedTransactionStorage extends DistributedTransa
   async save(
     key: string,
     data: TransactionCheckpoint,
-    ttl?: number
+    ttl?: number,
+    options?: TransactionOptions,
   ): Promise<void> {
-    this.storage.set(key, data)
-  }
+    const hasFinished = [
+      TransactionState.DONE,
+      TransactionState.REVERTED,
+      TransactionState.FAILED,
+    ].includes(data.flow.state)
 
-  async delete(key: string): Promise<void> {
-    this.storage.delete(key)
-  }
-
-  async archive(key: string, options?: TransactionModelOptions): Promise<void> {
-    this.storage.delete(key)
+    if (hasFinished) {
+      this.storage.delete(key)
+    } else {
+      this.storage.set(key, data)
+    }
   }
 }
